Make reflowing long lines optional
I had read that subtitle lines should not extend very long and made a little function that automatically wraps lines. This is a nice function, but most users will probably want to opt out. I've added a checkbox in the export window that will let users opt in if they want.